Split core into discrete files.
