+++ pkg/shipshape/result_test.go
@@ -1,0 +1,1 @@
+package shipshape_test
